fix(angular): clean ngc output before build to drop stale registry shadow
