+++ packages/navigation/package.json
@@ -151,7 +151,8 @@
   },
   "scripts": {
     "prepare": "node ../../scripts/vendor-codegen-specs.cjs react-native-screens src/fabric && pnpm run ng:build",
-    "ng:build": "ngc -p tsconfig.angular.json",
+    "clean": "rm -rf build",
+    "ng:build": "pnpm run clean && ngc -p tsconfig.angular.json",
     "typecheck": "tsc --build",
     "format": "prettier --write \"src/**/*.{ts,tsx}\""
   },
