+++ packages/slider/package.json
@@ -76,7 +76,8 @@
   "scripts": {
     "prepare": "node ../../scripts/vendor-codegen-specs.cjs @react-native-community/slider src && pnpm run ng:build",
     "typecheck": "tsc --build",
-    "ng:build": "ngc -p tsconfig.angular.json",
+    "clean": "rm -rf build",
+    "ng:build": "pnpm run clean && ngc -p tsconfig.angular.json",
     "format": "prettier --write \"src/**/*.{ts,tsx}\""
   },
   "dependencies": {
