+++ packages/splash-screen/package.json
@@ -75,7 +75,8 @@
   "scripts": {
     "prepare": "node ../../scripts/vendor-codegen-specs.cjs react-native-bootsplash src/specs && pnpm run ng:build",
     "typecheck": "tsc --build",
-    "ng:build": "ngc -p tsconfig.angular.json",
+    "clean": "rm -rf build",
+    "ng:build": "pnpm run clean && ngc -p tsconfig.angular.json",
     "format": "prettier --write \"src/**/*.{ts,tsx}\""
   },
   "dependencies": {
